Split Node.js into docker services "webapp" and "dev"
Service "webapp" is starting next.js and keeps running. This is similiar
to a production deployment

Service "dev" is for attaching to docker container bash and developing
from inside. It has several developer tools and configs for eg. git, vim, etc. already installed.

Add description about it in Readme.md