Update README with connection to db via psql
For newbies <-- it would be great to have a clear example of how to connect to the database and see the schema.
This is essential I would say to later than see how the impact of a change in `current.sql` will take immediate effect (i.e. migration happening in realtime).